Prepare the UI to show the 'domain filtered' spec file
Follow the steps shown in:
- [this diagram]()

This is to ensure we show the user the preview
of the file containing only 1 domain.

It may be the case of the user uploading an
already valid file.

But it can obly be the case of the user uploading
a file with several domains, so always show the
final version (the valis one).